commit postponed changes from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-3-1-branch@66121 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/include/R_ext/GraphicsDevice.h
@@ -850,7 +850,7 @@ LibExtern Rboolean mbcslocale;
 #endif
 
 /* Useful for devices: translates Adobe symbol encoding to UTF-8 */
-extern void *AdobeSymbol2utf8(char*out, const char *in, int nwork);
+extern void *AdobeSymbol2utf8(char*out, const char *in, size_t nwork);
 /* Translates Unicode point to UTF-8 */
 extern size_t Rf_ucstoutf8(char *s, const unsigned int c);
